confirm account

+++ frontend/expense-calculator/src/components/Login/Login.js
@@ -39,9 +39,13 @@ export default class Login extends React.Component {
                 ) 
             .then(res => res.json())
             .then(d=>{
+               if(!d.confirmed){
+                   alert('Please confirm your account')
+               }else {
                 localStorage.setItem('jwt', d.jwt);
                 localStorage.setItem('fullname', d.fullname);
-                window.location.reload()                
+                window.location.reload()      
+            }         
             })
             .catch(err =>{
                 console.log(err);
